Move version.py to head of project
